requirements.txt
absl-py==1.3.0
antlr4-python3-runtime==4.8
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
future==0.18.2
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
idna==3.4
importlib-metadata==5.0.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
mediapipe==0.8.11
MouseInfo==0.1.3
nltk==3.2.4
numpy==1.22.1
oauthlib==3.2.2
omegaconf==2.1.1
opencv-contrib-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.4.0
Pillow==9.1.1
preprocessing==0.1.13
protobuf==3.20.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pyDeprecate==0.3.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
seaborn==0.11.2
six==1.16.0
sphinx-rtd-theme==0.2.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.11.0
torchmetrics==0.8.0
torchvision==0.12.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
vgamepad==0.0.8
Werkzeug==2.2.2
zipp==3.10.0
